+++ lib/active_model/serializer/attributes.rb
@@ -34,10 +34,13 @@ module ActiveModel
         #   class AdminAuthorSerializer < ActiveModel::Serializer
         #     attributes :id, :name, :recent_edits
         def attributes(*attrs)
+          options = attrs.last.class == Hash ? attrs.pop : {}
+          options = options.except(:key)
+
           attrs = attrs.first if attrs.first.class == Array
 
           attrs.each do |attr|
-            attribute(attr)
+            attribute(attr, options)
           end
         end